refactor code: move realms prp sync to partners
