+++ src/etools/applications/partners/tasks.py
@@ -11,6 +11,7 @@ from django.utils.translation import gettext as _
 
 from celery.utils.log import get_task_logger
 from django_tenants.utils import get_tenant_model, schema_context
+from requests import HTTPError
 from unicef_snapshot.models import Activity
 from unicef_vision.exceptions import VisionException
 from unicef_vision.utils import get_data_from_insight
@@ -18,7 +19,10 @@ from unicef_vision.utils import get_data_from_insight
 from etools.applications.environment.notifications import send_notification_with_template
 from etools.applications.partners.models import Agreement, Intervention, PartnerOrganization
 from etools.applications.partners.prp_api import PRPAPI
-from etools.applications.partners.serializers.prp_v1 import PRPPartnerOrganizationWithStaffMembersSerializer
+from etools.applications.partners.serializers.prp_v1 import (
+    PRPPartnerOrganizationWithStaffMembersSerializer,
+    PRPSyncUserSerializer,
+)
 from etools.applications.partners.utils import (
     copy_all_attachments,
     send_intervention_draft_notification,
@@ -30,7 +34,7 @@ from etools.applications.partners.utils import (
 from etools.applications.partners.validation.agreements import AgreementValid
 from etools.applications.partners.validation.interventions import InterventionValid
 from etools.applications.reports.models import CountryProgramme
-from etools.applications.users.models import Country, User
+from etools.applications.users.models import Country, Realm, User
 from etools.config.celery import app
 from etools.libraries.djangolib.utils import get_environment
 from etools.libraries.tenant_support.utils import every_country, run_on_all_tenants
@@ -580,3 +584,30 @@ def send_pd_to_vision(tenant_name: str, intervention_pk: int, retry_counter=0):
                              f'PD number: {intervention_pk}. Business area code: {tenant.business_area_code}')
     finally:
         connection.set_tenant(original_tenant)
+
+
+@app.task
+def sync_realms_to_prp(user_pk, last_modified_at, retry_counter=0):
+    last_modified_instance = Realm.objects.filter(user_id=user_pk).order_by('modified').last()
+    if last_modified_instance and last_modified_instance.modified != last_modified_at:
+        # there were updates to user realms. skip
+        return
+
+    user = User.objects.filter(pk=user_pk).prefetch_related(
+        Prefetch('realms', Realm.objects.filter(is_active=True).select_related('country', 'organization', 'group')),
+    ).get()
+    data = PRPSyncUserSerializer(instance=user).data
+
+    try:
+        PRPAPI().send_user_realms(data)
+    except HTTPError as ex:
+        if retry_counter < 2:
+            logger.info(f'Received {ex} from prp api. retrying')
+            sync_realms_to_prp.apply_async(
+                (user_pk, last_modified_at),
+                {'retry_counter': retry_counter + 1},
+                eta=timezone.now() + datetime.timedelta(minutes=1 + retry_counter)
+            )
+        else:
+            logger.exception(f'Received {ex} from prp api while trying to send realms after 3 attempts. '
+                             f'User pk: {user_pk}.')
