--- src/etools/applications/users/apps.py
@@ -1,9 +1,0 @@
-from django.apps import AppConfig as BaseAppConfig
-
-
-class AppConfig(BaseAppConfig):
-    name = __name__.rpartition('.')[0]
-    verbose_name = 'Users'
-
-    def ready(self):
-        from . import signals  # NOQA
