+++ src/etools/applications/users/serializers.py
@@ -268,28 +268,3 @@ class CountrySerializer(SimpleUserSerializer):
             'iso3_code',
             'schema_name',
         )
-
-
-class PRPSyncRealmSerializer(serializers.ModelSerializer):
-    country = serializers.CharField(source='country.id')
-    organization = serializers.CharField(source='organization.vendor_number')
-    group = serializers.CharField(source='group.name')
-
-    class Meta:
-        model = Realm
-        fields = (
-            'country',
-            'organization',
-            'group',
-        )
-
-
-class PRPSyncUserSerializer(serializers.ModelSerializer):
-    realms = PRPSyncRealmSerializer(many=True)
-
-    class Meta:
-        model = get_user_model()
-        fields = (
-            'email',
-            'realms',
-        )
